chore: move all example plugins to a common directory
