+++ examples/plugins/echo/echo.go
