+++ examples/plugins/echo/locales/en/en.yaml
